Last tries with the VPU driver
Since the #rockchip channel is completely useless, my remaining
hopes are the Rockchip mainline maintainers.

Anyway, here's Ezequiel Garcia code for the VPU JPEG encoder.
It works and is useable.

I'll use it as a basis (again) for trying to decode H264 using
the test data provided by the bootlin guys !

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>