Fuck forge interoperability.
They don't give a shit about that, anyway, don't they ?

Whoever interprets /blah as a relative link is a moron.
[RockMyy](/Miouyouyou/RockMyy) is supposed to link to

* https://github.com/Miouyouyou/RockMyy ;
* https://gitlab.com/Miouyouyou/RockMyy ;
* https://notabug.com/Miouyouyou/RockMyy ;

depending on the current host.

Turns out that none of these forges interpret it as
such. They interpret it as
[current_url_of_README.md]/Miouyouyou/RockMyy

Seriously, I'm not going to write one README per forge.

So I'll put Github links in the README, since it's the
most used forge at the moment.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>